pear_email_ingest_mvp_imap/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
google-cloud-storage==2.17.0
google-cloud-aiplatform==1.70.0
mysql-connector-python==9.0.0
pydantic==2.8.2
